requirements.txt
flask
clova-cek-sdk
line-bot-sdk
